Fixed log verification test and moved test to new folder to keep consistent structure.
